+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## [Unreleased]
+
+### Added
+
+- concord-server: `afterCreatedAt` parameter to the process list API
+endpoint.
+
+
+
 ## [0.94.0] - 2018-10-29
 
 ### Added